// src/declarations/file_manager/file_manager.did.d.ts
export interface FileResponse {
  id: bigint;
  permission: Permission;
  updated_at: bigint;
  sha256: Uint8Array | number[];
  owner: [] | [Principal];
  metadata: [] | [string];
  name: string;
  path: string;
  size: bigint;
  mime_type: string;
  created_at: bigint;
  parent_id: [] | [bigint];
  chunks: BigUint64Array | bigint[];
  is_protected: boolean;
  extension: string;
}
// ...
export type Permission = { Private: null } | { Origin: Array<string> } | { Public: null };